Refactor: ProductJson reading to dedicated unit
